finish tests and get foldes in order
