+++ .eslintrc.js
@@ -18,7 +18,8 @@ module.exports = {
         "react/jsx-uses-vars": [2],
         "indent": [
             "error",
-            2
+            2, 
+            { "SwitchCase": 1 }
         ],
         "linebreak-style": [
             "error",
