+++ jsx/lighting/channel.jsx
@@ -1,186 +1,181 @@
 import React from 'react'
-import { HuePicker } from 'react-color'
-import Profile from 'lighting/profile'
-import PropTypes from 'prop-types'
+import ColorPicker from './color_picker'
+import ProfileSelector from './profile_selector'
+import Profile from './profile'
+import Percent from './percent'
+import {ErrorFor, NameFor, ShowError, PathToObject} from '../utils/validation_helper'
+import {Field} from 'formik'
 
-export default class Channel extends React.Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      channel: this.props.config,
-      colorPicked: false,
-      expand: false
+const Channel = (props) => {
+
+  const handleChange = e => {
+    props.onChangeHandler(e, props.channelNum)
+  }
+
+  const configFor = profileType => {
+    let touched = props.touched
+    let config = PathToObject(props.name + '.profile.config', touched)
+
+    switch(profileType){
+      case 'diurnal': {      
+        if (config){
+          config = {
+            start: false,
+            end: false
+          }
+          props.setTouched(touched)      
+        }
+
+        return {
+          start: '',
+          end: ''
+        }
+      }
+      case 'fixed': {
+        if (config){
+          config = { value: false }
+          props.setTouched(touched)
+        }
+
+        return {value: 0}
+      }
+      case 'auto': {
+        if (config){
+          config = { values: false }
+          props.setTouched(touched)
+        }
+
+        return {values: [0,0,0,0,0,0,0,0,0,0,0,0]}
+      }
+      default: 
+        return {}
     }
-    this.updateReverse = this.updateReverse.bind(this)
-    this.update = this.update.bind(this)
-    this.updateMin = this.updateMin.bind(this)
-    this.updateMax = this.updateMax.bind(this)
-    this.updateStartMin = this.updateStartMin.bind(this)
-    this.updateName = this.updateName.bind(this)
-    this.updateColor = this.updateColor.bind(this)
-    this.colorPicker = this.colorPicker.bind(this)
-    this.toggle = this.toggle.bind(this)
-    this.details = this.details.bind(this)
   }
 
-  toggle () {
-    this.setState({expand: !this.state.expand})
-  }
-
-  colorPicker () {
-    if (!this.state.colorPicked) {
-      return (
-        <button
-          disabled={this.props.readOnly}
-          onClick={() => this.setState({colorPicked: true})}
-          style={{backgroundColor: this.state.channel.color}}
-          className='btn btn-secondary'
-        />
-      )
+  const handleConfigChange = e => {
+    console.log('config changing')
+    const event = {
+      target: {
+        name: e.target.name,
+        value: {
+          type: e.target.value,
+          config: configFor(e.target.value)
+        }
+      }
     }
-    return (
-      <HuePicker color={this.state.channel.color} onChangeComplete={this.updateColor} />
-    )
+    props.onChangeHandler(event, props.channelNum)
   }
 
-  updateMin (ev) {
-    this.update('min', ev.target.value)
-  }
 
-  updateColor (color) {
-    var ch = this.state.channel
-    ch.color = color.hex
-    this.setState({
-      channel: ch,
-      colorPicked: false
-    })
-    this.props.hook(ch)
-  }
+  return (
+    <div className="controls border-top">
+      <div className="row align-items-start">
+        <div className="col-sm-6 col-md-4 col-xl-2">
+          <div className="form-group">
+            <label className="w-100">
+              Name 
+              <small className="float-right badge badge-info mt-1">(pin {props.channel.pin})</small>
+            </label>
+            <Field name={NameFor(props, 'name')}
+              className={ShowError(props, NameFor(props,'name')) ? 'form-control is-invalid' : 'form-control'} 
+              placeholder="Channel Name"
+              disabled={props.readOnly} />            
+            <ErrorFor {...props}
+              name={NameFor(props, 'name')}  />
+          </div>          
+        </div>
+        
+        <div className="form-group col-sm-6 col-md-4 col-xl-2 form-inline">
+          <label className="mb-2">Color</label>
+          <ColorPicker name={NameFor(props, 'color')}
+            readOnly={props.readOnly}
+            color={props.channel.color}
+            onChangeHandler={handleChange} />
+        </div>
 
-  updateName (ev) {
-    this.update('name', ev.target.value)
-  }
-
-  updateStartMin (ev) {
-    this.update('start_min', ev.target.value)
-  }
-
-  updateMax (ev) {
-    this.update('max', ev.target.value)
-  }
-
-  update (k, v) {
-    var ch = this.state.channel
-    ch[k] = v
-    this.setState({
-      channel: ch
-    })
-    this.props.hook(ch)
-  }
-
-  updateReverse (ev) {
-    this.update('reverse', ev.target.checked)
-  }
-
-  details () {
-    return (
-      <div className='container border-light'>
-        <div className='form-inline row'>
-          <div className='form-group col-lg-2'>
-            <label>Reverse</label>
-            <input
-              disabled={this.props.readOnly}
-              type='checkbox'
-              onClick={this.updateReverse}
-              defaultChecked={this.state.channel.reverse}
-              className='form-control'
-            />
+        <div className="col-sm-6 col-md-4 col-xl-2">
+          <div className="form-group">
+            <label>Behavior</label>
+            <select className="custom-select"
+              name={NameFor(props, 'reverse')}
+              disabled={props.readOnly}
+              onChange={handleChange}
+              onBlur={props.onBlur}
+              value={props.channel.reverse}>
+              <option value="false">Active High</option>
+              <option value="true">Active Low</option>
+            </select>
           </div>
-          <div className='form-group col-lg-2'>
-            <label>Color</label>
-            {this.colorPicker()}
-          </div>
-          <div className='form-group col-lg-2'>
+        </div>
+        <div className="col-sm-6 col-md-4 col-xl-2">
+          <div className="form-group">
             <label>Min</label>
-            <input
-              type='text'
-              onChange={this.updateMin}
-              value={this.state.channel.min}
-              className='form-control col-lg-6'
-              disabled={this.props.readOnly}
-            />
-          </div>
-          <div className='form-group col-lg-2'>
+            <Percent type="text" 
+              className={ShowError(props, NameFor(props, 'min')) ? 'form-control is-invalid' : 'form-control'} 
+              name={NameFor(props, 'min')} 
+              onBlur={props.onBlur}
+              disabled={props.readOnly}
+              onChange={handleChange}
+              value={props.channel.min} />
+            <ErrorFor {...props} name={NameFor(props, 'min')} />
+          </div>  
+        </div>
+        <div className="col-sm-6 col-md-4 col-xl-2">
+          <div className="form-group">
             <label>Max</label>
-            <input
-              className='form-control col-lg-6'
-              type='text'
-              onChange={this.updateMax}
-              value={this.state.channel.max}
-              disabled={this.props.readOnly}
-            />
-          </div>
-          <div className='form-group col-lg-2'>
+            <Percent type="text" 
+              className={ShowError(props, NameFor(props, 'max')) ? 'form-control is-invalid' : 'form-control'} 
+              name={NameFor(props, 'max')}
+              onBlur={props.onBlur}
+              disabled={props.readOnly}
+              onChange={handleChange}
+              value={props.channel.max}  />
+            <ErrorFor {...props} name={NameFor(props, 'max')} />
+          </div>  
+        </div>
+        <div className="col-sm-6 col-md-4 col-xl-2">
+          <div className="form-group">
             <label>Start</label>
-            <input
-              className='form-control col-lg-6'
-              type='text'
-              onChange={this.updateStartMin}
-              value={this.state.channel.start_min}
-              disabled={this.props.readOnly}
-            />
-          </div>
-        </div>
-        <div className='row'>
-          <Profile
-            config={this.state.channel.profile.config}
-            type={this.state.channel.profile.type}
-            hook={(v) => this.update('profile', v)}
-            name={'profile-' + this.props.light_id + '-' + this.state.channel.pin}
-            readOnly={this.props.readOnly}
-          />
+            <Percent type="text" 
+              className={ShowError(props, NameFor(props, 'start_min')) ? 'form-control is-invalid' : 'form-control'} 
+              name={NameFor(props, 'start_min')}
+              onBlur={props.onBlur}
+              disabled={props.readOnly}
+              onChange={handleChange} 
+              value={props.channel.start_min} />
+            <ErrorFor {...props} name={NameFor(props, 'start_min')} />
+          </div>  
         </div>
       </div>
-    )
-  }
-
-  render () {
-    var d = <div />
-    var lbl = '+'
-    if (this.state.expand) {
-      d = this.details()
-      lbl = '-'
-    }
-    return (
-      <div className='container border-light rounded'>
-        <div className='form-inline'>
-          <div className='form-group'>
-            <input
-              type='text'
-              value={this.state.channel.name}
-              onChange={this.updateName}
-              className='form-control'
-              disabled={this.props.readOnly}
-            />
-            <span className='badge badge-secondary'>({this.state.channel.pin})</span>
-          </div>
-          <input
-            type='button'
-            value={lbl}
-            className='btn btn-secondary'
-            onClick={this.toggle}
-            id={'expand-channel-' + this.props.light_id + '-' + this.state.channel.pin}
-          />
+      <div className="row">
+        <div className="col">
+          <div className="form-group">
+            <label className="mr-3">Profile</label>
+            <ProfileSelector
+              className={ShowError(props, NameFor(props, 'profile.type')) ? 'form-control is-invalid' : 'form-control'} 
+              name={NameFor(props, 'profile')}
+              readOnly={props.readOnly}
+              onChangeHandler={handleConfigChange}
+              value={props.channel.profile.type} />
+            <input className="d-none is-invalid form-control"/> 
+            <ErrorFor {...props} name={NameFor(props, 'profile.type')} />
+          </div>          
         </div>
-        {d}
       </div>
-    )
-  }
+      <div className="row mb-3">
+        <div className="col">
+          <Profile 
+            {...props}
+            name={NameFor(props, 'profile.config')}
+            onBlur={props.onBlur}
+            readOnly={props.readOnly}
+            type={props.channel.profile.type}
+            value={props.channel.profile.config} 
+            onChangeHandler={handleChange} />
+        </div>
+      </div>
+    </div>
+  )
 }
 
-Profile.propTypes = {
-  light_id: PropTypes.string,
-  hook: PropTypes.func,
-  config: PropTypes.object,
-  readOnly: PropTypes.bool
-}
+export default Channel
