+++ jsx/lighting/diurnal_profile.jsx
@@ -1,94 +1,40 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Formik, Field } from 'formik'
-import * as Yup from 'yup'
-import FormikObserver from 'utils/form_observer'
+import {ErrorFor, NameFor, ShowError} from 'utils/validation_helper'
+import {Field} from 'formik'
 
-const diurnalSchema = Yup.object().shape({
-  start: Yup.string()
-    .required('Start time is required.')
-    .matches(/^([01]?[0-9]|2[0-3]):[0-5][0-9]$/, 'Start time must be a valid time (HH:MM).'),
-  end: Yup.string()
-    .required('End time is required.')
-    .matches(/^([01]?[0-9]|2[0-3]):[0-5][0-9]$/, 'End time must be a valid time (HH:MM).')
-})
+const DiurnalProfile = (props) => {
 
-export default class DiurnalProfile extends React.Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      config: {
-        start: props.config && props.config.start ? props.config.start : '',
-        end: props.config && props.config.end ? props.config.end : ''
-      }
-    }
-    this.update = this.update.bind(this)
-  }
-
-  update (k) {
-    var s = this.state.config
-    return (ev) => {
-      console.log('Updating ')
-      s[k] = ev.target.value
-      this.props.hook(s)
-    }
-  }
-
-  render () {
-    return (
-      <Formik
-        initialValues={{
-          start: this.state.config.start,
-          end: this.state.config.end
-        }}
-        validationSchema={diurnalSchema}
-        render={({values, errors, touched, handleChange, handleBlur}) => (
-          <form>
-            <div className='form-inline row align-items-start'>
-              <div className='form-group col-lg-4'>
-                <label className='col-form-label col-sm-5'>Start Time</label>
-                <input
-                  type='text'
-                  name='start'
-                  onBlur={handleBlur}
-                  onChange={handleChange}
-                  readOnly={this.props.readOnly}
-                  className={'form-control col-lg-6 ' + (errors.start && touched.start ? 'is-invalid' : '')}
-                  value={values.start}
-                />
-                {
-                  errors.start && touched.start && 
-                  (
-                    <div className="field-error invalid-feedback text-center">{errors.start}</div>
-                  )
-                }
-              </div>
-              <div className='form-group col-lg-5'>
-                <label className='col-form-label col-sm-5'>End Time</label>
-                <Field type='text' name='end' required
-                  className={'form-control col-lg-6 ' + (errors.end && touched.end ? 'is-invalid' : '')}
-                  readOnly={this.props.readOnly}
-                />
-                {
-                  errors.end && touched.end && 
-                  (
-                    <div className="field-error invalid-feedback text-center">{errors.end}</div>
-                  )
-                }
-              </div>
-              <FormikObserver
-                onChange={({ values }) => console.log('onchange', values)}
-              />
-            </div>
-          </form>
-        )} 
+  return (
+    <div className="form-inline">
+      
+      <label className='mr-2'>Start Time</label>
+      <Field name={NameFor(props, 'start')}
+        readOnly={props.readOnly}
+        className={ShowError(props, NameFor(props, 'start')) ? 'form-control mr-3 col-12 col-sm-3 col-md-2 col-lg-2 is-invalid' : 'form-control mr-3 col-12 col-sm-3 col-md-2 col-lg-2'}
+        placeholder="HH:mm"
+      />      
+      <label className='mr-2'>End Time</label>
+      <Field name={NameFor(props, 'end')}
+        readOnly={props.readOnly}
+        className={ShowError(props, NameFor(props, 'end')) ? 'form-control col-12 col-sm-3 col-md-2 col-lg-2 is-invalid' : 'form-control col-12 col-sm-3 col-md-2 col-lg-2'}
+        placeholder="HH:mm"
       />
-    )
-  }
+      <ErrorFor {...props} name={NameFor(props, 'start')} />
+      <ErrorFor {...props} name={NameFor(props, 'end')} />
+    </div>
+  )
+}
+
+DiurnalProfile.defaultProps = {
+  start: '',
+  end: ''
 }
 
 DiurnalProfile.propTypes = {
   config: PropTypes.object,
-  hook: PropTypes.func,
-  readOnly: PropTypes.bool
+  readOnly: PropTypes.bool,
+  onChangeHandler: PropTypes.func.isRequired
 }
+
+export default DiurnalProfile
