+++ jsx/lighting/main.jsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import $ from 'jquery'
-import Light from '../lights/light_form'
+import Light from './light_form'
 import {showAlert} from 'utils/alert'
-import {confirm} from 'utils/confirm'
 import {updateLight, fetchLights, createLight, deleteLight} from 'redux/actions/lights'
 import {fetchJacks} from 'redux/actions/jacks'
 import {connect} from 'react-redux'
